Fix upside-down vflip bit images
For some unknown reason, I used x-origin and y-origin instead of checking the correct bits in the image descriptor.
THIS IS NOT how you determine if the origin is bottom-left or top-left.

Also, it will be better not to flip an image after loading in load_tga_ext(). Better to leave that task to the person
loading it for optimization, as OpenGL, for example, expects bottom-left images.